// src/app/pages/eventos/Model/eventoAdd.interface.ts
export interface EventoAddRequest {
    idInmueble: number;
    idEstatusEvento: number;
    fechaEvento: string;
    reciboNumero: number;
    fechaPago: string;
    apartado: number;
    liquida: number;
    luz: number;
    depositoGarantia: number;
    limpiezaDomingo: number;
    rentaInmobiliario: number;
}